etls/myfirstETL/pyproject.toml
[tool.poetry]
name = "myfirstetl"
version = "0.1.0"
description = "First ETL with pandas and scikit-learn"
authors = ["David Castillo <dcastilloa@uni.pe>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
scikit-learn = "^1.5.0"
google-cloud-storage = "^2.18.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
poethepoet = "^0.29.0"
pytest = "^8.3.0"

[tool.poe.tasks]
ruff.shell = "ruff check src/ tests/ --fix && ruff format src/ tests/"
run = "python src/pipeline.py"
test = "pytest tests/"

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true
exclude = [".git", ".venv", "__pycache__", "*.egg-info"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
